% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_block_20_race_df}
\alias{nyc_block_20_race_df}
\title{New York City race and Hispanic origin by Census block, 2020}
\format{
A tibble with 37,984 rows and 25 columns:
\describe{
\item{geoid}{15-character block GEOID (state + county + tract + block).}
\item{county}{NYC county (borough) name derived from \code{geoid}.}
\item{name}{Census block label as returned by the Census Bureau.}
\item{total_pop}{Total population of the block (PL summary variable
\code{P1_001N}).}
\item{white_alone}{Population reporting White alone (\code{P1_003N}).}
\item{black_alone}{Population reporting Black or African American alone
(\code{P1_004N}).}
\item{aian_alone}{American Indian and Alaska Native alone (\code{P1_005N}).}
\item{asian_alone}{Asian alone (\code{P1_006N}).}
\item{hipi_alone}{Native Hawaiian and Other Pacific Islander alone
(\code{P1_007N}).}
\item{other_alone}{Some other race alone (\code{P1_008N}).}
\item{two_or_more}{Two or more races (\code{P1_009N}).}
\item{hispanic}{Hispanic or Latino population (\code{P2_002N}).}
\item{non_hispanic}{Not Hispanic or Latino population (\code{P2_003N}).}
\item{nh_white_alone}{Not Hispanic, White alone (\code{P2_005N}).}
\item{nh_black_alone}{Not Hispanic, Black or African American alone
(\code{P2_006N}).}
\item{nh_aian_alone}{Not Hispanic, AIAN alone (\code{P2_007N}).}
\item{nh_asian_alone}{Not Hispanic, Asian alone (\code{P2_008N}).}
\item{nh_hipi_alone}{Not Hispanic, NHOPI alone (\code{P2_009N}).}
\item{nh_other_alone}{Not Hispanic, some other race alone (\code{P2_010N}).}
\item{nh_two_or_more}{Not Hispanic, two or more races (\code{P2_011N}).}
\item{nh_white_prop}{Proportion of \code{total_pop} that is non-Hispanic White
alone. \code{NA} when \code{total_pop == 0}.}
\item{nh_black_prop}{Proportion non-Hispanic Black alone.}
\item{nh_asian_prop}{Proportion non-Hispanic Asian alone.}
\item{hispanic_prop}{Proportion Hispanic or Latino.}
\item{nh_two_or_more_prop}{Proportion non-Hispanic two or more races.}
}
}
\source{
US Census Bureau, 2020 Decennial Census, PL 94-171 Redistricting
Data Summary File. Retrieved via \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}.
\url{https://www.census.gov/programs-surveys/decennial-census/about/rdo/summary-files.html}
}
\usage{
nyc_block_20_race_df
}
\description{
Population counts for the five New York City counties at the 2020 Census
block level, broken down by race and Hispanic origin. Sourced from the
2020 Decennial Census PL 94-171 redistricting summary file. Joins to
\link[nycmaps:nyc_census_blocks_2020_sf]{nycmaps::nyc_census_blocks_2020_sf} on \code{geoid}.
}
\keyword{datasets}
