% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyc_counties.R
\name{nyc_counties}
\alias{nyc_counties}
\title{New York City county names}
\usage{
nyc_counties()
}
\value{
A character vector of length 5: \code{"New York"}, \code{"Kings"},
\code{"Queens"}, \code{"Bronx"}, \code{"Richmond"}.
}
\description{
The five counties (boroughs) that make up New York City, in the form
expected by \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}} and \code{\link[tidycensus:get_decennial]{tidycensus::get_decennial()}}
(i.e. county \emph{names}, not FIPS codes).
}
\examples{
nyc_counties()

}
