% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_county_acs_race_df}
\alias{nyc_county_acs_race_df}
\title{New York City race and Hispanic origin by county, ACS 5-year}
\format{
A tibble with 5 rows and 25 columns:
\describe{
\item{geoid}{5-character county GEOID.}
\item{county}{NYC county (borough) name derived from \code{geoid}.}
\item{name}{County label as returned by the Census Bureau.}
\item{total_pop}{Total population (\code{B03002_001} summary estimate).}
\item{total_pop_moe}{Margin of error for \code{total_pop}.}
\item{nh_white}{Not Hispanic, White alone (\code{B03002_003}).}
\item{nh_black}{Not Hispanic, Black or African American alone
(\code{B03002_004}).}
\item{nh_aian}{Not Hispanic, American Indian and Alaska Native alone
(\code{B03002_005}).}
\item{nh_asian}{Not Hispanic, Asian alone (\code{B03002_006}).}
\item{nh_hipi}{Not Hispanic, Native Hawaiian and Other Pacific Islander
alone (\code{B03002_007}).}
\item{nh_other}{Not Hispanic, some other race alone (\code{B03002_008}).}
\item{nh_two_or_more}{Not Hispanic, two or more races (\code{B03002_009}).}
\item{hispanic}{Hispanic or Latino, any race (\code{B03002_012}).}
\item{nh_white_moe, nh_black_moe, nh_aian_moe, nh_asian_moe, nh_hipi_moe,
nh_other_moe, nh_two_or_more_moe, hispanic_moe}{Margins of error for
the corresponding estimates.}
\item{nh_white_prop}{Proportion non-Hispanic White. \code{NA} when
\code{total_pop == 0}.}
\item{nh_black_prop}{Proportion non-Hispanic Black.}
\item{nh_asian_prop}{Proportion non-Hispanic Asian.}
\item{hispanic_prop}{Proportion Hispanic or Latino.}
}
}
\source{
US Census Bureau, 2020-2024 American Community Survey 5-year
estimates, table B03002. Retrieved via \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}.
\url{https://www.census.gov/programs-surveys/acs}
}
\usage{
nyc_county_acs_race_df
}
\description{
Race and Hispanic-origin estimates from the 2020-2024 American Community
Survey 5-year release for the five New York City counties (boroughs).
Sourced from ACS table B03002 ("Hispanic or Latino origin by race"), with
\code{B03002_001} as the population denominator. Each estimate is accompanied
by its 90\% margin of error, with the suffix \verb{_moe}. The county-level
counterpart of \link{nyc_tract_acs_race_df}. Joins to \link[nycmaps:nyc_boros_sf]{nycmaps::nyc_boros_sf}
via the \code{county} column, which matches
\code{nycmaps::nyc_boros$short_county_name}.
}
\keyword{datasets}
