% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nyc_tract_acs_education_df}
\alias{nyc_tract_acs_education_df}
\title{New York City educational attainment by Census tract, ACS 5-year}
\format{
A tibble with 2,327 rows and 26 columns:
\describe{
\item{geoid}{11-character tract GEOID.}
\item{county}{NYC county (borough) name derived from \code{geoid}.}
\item{name}{Census tract label as returned by the Census Bureau.}
\item{edu_total}{Population 25 years and over (\code{B15003_001}).}
\item{edu_less_than_hs}{No schooling completed (\code{B15003_002}).}
\item{edu_some_hs}{9th to 12th grade, no diploma (\code{B15003_016}).}
\item{edu_hs}{Regular high school diploma (\code{B15003_017}).}
\item{edu_some_college}{Some college, 1 or more years, no degree
(\code{B15003_019}).}
\item{edu_associates}{Associate's degree (\code{B15003_021}).}
\item{edu_bachelors}{Bachelor's degree (\code{B15003_022}).}
\item{edu_masters}{Master's degree (\code{B15003_023}).}
\item{edu_professional}{Professional school degree (\code{B15003_024}).}
\item{edu_doctorate}{Doctorate degree (\code{B15003_025}).}
\item{edu_total_moe, edu_less_than_hs_moe, edu_some_hs_moe, edu_hs_moe,
edu_some_college_moe, edu_associates_moe, edu_bachelors_moe,
edu_masters_moe, edu_professional_moe, edu_doctorate_moe}{Margins of
error for the corresponding estimates.}
\item{bachelors_or_higher}{\code{edu_bachelors + edu_masters + edu_professional + edu_doctorate}.}
\item{bachelors_or_higher_prop}{Proportion of \code{edu_total} with a
bachelor's degree or higher. \code{NA} when \code{edu_total == 0}.}
\item{less_than_hs_prop}{Proportion of \code{edu_total} whose highest
attainment is below a high school diploma.}
}
}
\source{
US Census Bureau, 2020-2024 American Community Survey 5-year
estimates, table B15003. Retrieved via \code{\link[tidycensus:get_acs]{tidycensus::get_acs()}}.
\url{https://www.census.gov/programs-surveys/acs}
}
\usage{
nyc_tract_acs_education_df
}
\description{
Educational attainment for the population 25 years and over from the
2020-2024 American Community Survey 5-year release for the five New York
City counties at the Census tract level. Sourced from ACS table B15003.
Each estimate is accompanied by its 90\% margin of error, with the suffix
\verb{_moe}. Joins to \link[nycmaps:nyc_census_tracts_2020_sf]{nycmaps::nyc_census_tracts_2020_sf} on \code{geoid}.
}
\keyword{datasets}
