% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycdemog-package.R
\docType{package}
\name{nycdemog-package}
\alias{nycdemog}
\alias{nycdemog-package}
\title{nycdemog: New York City Demographic Data from the Census}
\description{
Tract- and block-level demographic data for the five counties of New York City, derived from the US Census Bureau via 'tidycensus'. Provides curated tibbles keyed by GEOID and thin wrapper functions for fetching custom NYC slices on demand.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kjhealy.github.io/nycdemog/}
  \item \url{https://github.com/kjhealy/nycdemog}
  \item Report bugs at \url{https://github.com/kjhealy/nycdemog/issues}
}

}
\author{
\strong{Maintainer}: Kieran Healy \email{kjhealy@gmail.com} (\href{https://orcid.org/0000-0001-9114-981X}{ORCID})

Authors:
\itemize{
  \item Kieran Healy \email{kjhealy@gmail.com} (\href{https://orcid.org/0000-0001-9114-981X}{ORCID})
}

}
\keyword{internal}
